Blacklist ets_plugin on crystal arm64 build
The ets_plugin package depends on a third party library that doesn't build for arm, making the ets_plugin build to fail: http://build.ros2.org/job/Cbin_ubv8_uBv8__ets_plugin__ubuntu_bionic_arm64__binary/
The package is a plugin for a game (Euro Truck Simulator) that doesn't support arm, so it actually doesn't make sense to try to build it for that architecture. Thus adding it to the package blacklist for the arm64 build seems to be the proper solution.